+++ code/source/OMIP.cpp
@@ -44,13 +44,13 @@ std::vector<double> OMIP::getSol(){
 }   
 
 void OMIP::setup(){
-    for(int i=0;i<getNumRows();i++){
+    for(size_t i=0;i<getNumRows();i++){
         std::vector<double> col(getNumRows(),0.);
         col[i]=1;
         addCol(col,OMIP::OMIP_SLACK_OBJ_COEFF,0.,CPX_INFBOUND,"SP_"+std::to_string(i+1));
     }
 
-    for(int i=0;i<getNumRows();i++){
+    for(size_t i=0;i<getNumRows();i++){
         std::vector<double> col(getNumRows(),0.);
         col[i]=-1;
         addCol(col,OMIP::OMIP_SLACK_OBJ_COEFF,0.,CPX_INFBOUND,"SN_"+std::to_string(i+1));
@@ -62,7 +62,7 @@ void OMIP::setup(){
 void OMIP::addBudgetConstr(double rhs){
     std::vector<double> budConstr(getNumCols(),0.);
     int start = (getNumCols()-2*getNumRows());
-    for(int i=start;i<getNumCols();i++){
+    for(size_t i=start;i<getNumCols();i++){
         budConstr[i]=1.;
     }
     addRow(budConstr,OMIP::OMIP_BUD_CONST_SENSE,rhs);
